Calculate averaged translation and rotation contributions
Completely reworked the foundation of the algorithm.
Now, translation and rotation are treated similarily, and rely on each other.

Moreover, we now treat each track individually, and calculate a real weighted average
Each track provides a normalised contribution, both for translation and rotation.

TODO: the initialisation of the baselines per track