Build compenstation from measured data
For the 2D-stabilisation to work properly, it is necessary that the
transformations to compensate the movement are built in accordance
with the way the measurements for the individual contributions are
retrieved.

Most notably, the rotation must be performed around the same center,
which was used to determine the angular contribution. Currently we
use the translation compensated center of frame as this pivot.

In terms of code organisation it is better to separate the *measurement*
from the construction of a compensating function. This means that the
translation parameter should not contain any scale adjustments, since
such adjustments are a detail of how the compensation is done, and
actually might be done in multiple ways.

TODO: I have yet to figure out how to build the processing chain
for the compositor nodes in a similar fashion. I doubt the default
rotation works correct for our case. Anoter unsolved problem is
how to get the animation running (it works when played back in
the clip editor, but not in the sequence editor)