New Feature: Support for Zoom / Scale stabilisation
After all the refacturing, this is low hanging fruit now. Since any rotation
is basically a linear transform, we can as well pick up the scale factor
relative to the rotation pivot point.

To be able to handle this scale information in the existing framework,
we determine the vector lenghts at the reference frame, plus we use the
averaged scale at that frame as baseline. The individual contributions are
recorded on a logarithmic scale, to allow caluclating weighted averages.