replace the trickery of auto-faded tracks by a real solution with interpolation
Beyond the area covered with usable tracking data, the stabilization function
will be disabled, causing the image to jump back to neutral position.
From a users POV this behaviour is surprising; rather you'd expect the image
just to stay as it was, i.e. we want the stabilization to stall in the end position
instead of jumping to zero.

Previously I had a bit of magic trickery built into the marker access function
(basically the original solution in blender used a similar approach at the same location)

But on second thought, I consider this non obvious magic and decided to go
for a real solution, which unfortunately is way more expensive in terms of code.

Now, in this reworked version, when we detect we're outside the definition range,
we actively search the nearest position still in definition range; then we
either hold that value (for extrapolation) or interpolate in case of a gap
in the middle of the definiton range.

NOTE: right now this solution is affected from the "automation problem"
in the same way as the autoscale function is: since we don't (yet) use the
correct automation values when we're measuring stabilization data for anything
*other* than the *current frame*, the interpolation/holding could produce
quite strange results. This problem will go away though, once we handle
automation values more correct in the calculation.