Bugfix: need to handle pivot point similar for measurement and compensation
Problem was how the pivot behaves relative to a given target position.
For the measurement, we kept the pivot anchored relative to the reference
frame. But the way the compensation matrix is built effectively means that
the pivot is shifted to the target position (this is also a plausible choice,
since the target position represents the movement the user *wants*, e.g. a pan;
it seems plausible the user wants the pivot for rotation to travel with the
pan -- if this assumption is indeed apropriate needs further discussion though).