Bugfix: really use closest enabled marker for track initialisation
Previously we relied on BKE_tracking_marker_get, which contains magic to
fall back on the "closest" (actually next to left) frame, in case there
isn't an exact match.

Unfortunately, when refining tracks interactively, Blender has the habit
to set an additional marker to the previous frame with flag == disabled,
in order to mark the start of an active track segment. Thus, if we don't
observe the enabled state of a marker, we might end up using some arbitrary
incorrect tracking values, which just happen to sit in such a toggle marker.

When this happens during initialisation, these arbitrary values can poison
every other measurement value on that track and any other tracks chained behind.

The fix here is to re-use the search functions implemented for interpolation
of full gaps. This gives us the nice additional benefit of really picking
the marker closest to the the desired reference frame