use animated F-curve values only when necessary
Using the global "current values", as placed by the animation system
into the data structures, has the benefit of immediately responsive
GUI controls. Thus we switch to fetching the values from F-curves
only when we know we're accessing other frames than the global
"current frame" -- that is, when initialising the track baseline values,
when interpolating over gaps and when calculating the auto-scale.