Bugfix: correct use of Blender's pixel_aspect
Found this out while working with anamorphotic HDV footage: I had
misinterpreted the meaning of Blender's pixel_aspect value. In fact,
the tracking works on the raw data grid and produces normalised offsets.
Thus we need to *compensate* for the pixel aspect at the point where
we're changing to real canvas coordinates, prior to performing any
angular calculations. For example, in the case of anamorphotic footage,
we need to stretch out to the full intended image width.