Review: remove any access to deprecated (obsolete) rot_track
The old stabilizer implementation allowed to set
a single additional track for rotation stabilization.

In the reworked implementation, this is replaced by a set of
tracks for rotation/scale stabilization. Those are managed similar
to the location-stabilization tracks: by marking them in the
track->flags

This changeset removes any remaining access to the now obsolete
DNA field rot_track. Moreover the "join_tracks" operator is now
changed to evaluate and handle the track->flag setting properly