UI naming as proposed by Sean Kennedy: "Stabilize Scale"
unfortunately there is still a confusion regarding the "scale factor".
Historically (from old version of stabilizator), the maximum autoscale factor
was indicated and set using the inverse of the factor actually applied.
Thus we still display this autoscale factor in the traditional way,
while we display the "expected zoom" in the correct and natural way.

This is also the reason why I prefer to leave the UI label as "expected zoom"