Review: abandon caching of init values and re-init stabilizer on every frame
since proper initialization depends on (weight) animation and setup of tracks,
ensuring consistency of cached init data turns out to be tricky, hard to maintain
and generally not worth the effort.

Thus, in an effort to tackle the design problems with private working data,
we've decided to abandon this caching and re-initialize on every frame.
This effectively turns our private working data into per-frame data,
which can be passed with a context struct on the call graph.

To put this decision into context:
 - the previously used approach with value caching forced us to manage
   allocations in a tricky way, scattered over various external operators
 - yet still, we continued to discover init and consistency bugs, some of which
   are impossible to fix in an airtight way.
 - and last but not least, in practice we're already performing this
   init sequence very frequently, after every UI display update or
   value tweak.

We don't expect any noticeable performance degradation.