Review: introduce per call private working data context
We continue to use a GHash to associate a work struct for every track.
But this association is now via the StabContext, and thus memeory management
is straight forward: we allocate and discard everything per call to
BKE_tracking_stabilization_data_get

With a debug build, the slowdown due to initialisation on every frame
is noticeable, yet not annoying. On consecutive playback of the timeline,
the global caching of video frames kicks in. Thus no real performance problem!