clean-up: should be done on migration only
with migration, stabilizion.scale becomes a first class persistent
property. Previously it was only used to cache autoscale.
Autoscale is now treated as special case of a "target scale"

The stabilisation code can deal with scale == 0.0,
Moreover, it is not possible, to enter scale := 0.0 via UI
thus no need to check this on every load.