[stab-fix] BUGFIX to mitigate problems with rotation due to jumping pivot
When a shot can not be covered with continuous tracks in entirety, but
must be  stabilised with several overlapping tracks covering only a small
time span, then, whenever a new track enters the mix, the overall wight
centre of the active tracks jumps to another position. This causes all
existing tracks to suddenly pick up a wrong rotation contribution.

Unfortunately, the problem is, we do not know the relation between old
and new pivot point, since our calculation does not actually integrate over
all frames, but rather approximates this integration.

So this fix is an attempt to mitigate the problem: ongoing tracks now just
continue to use the pivot point they used from start, which is wrong to
the degree the point and the new pivot has drifted away, yet avoids
to calculate an outright wrong contribution with a sudden jump.