[stab-fix] revert to the simple solution
This mostly reverts the commit
"attempt to get a better approximation when changing pivot"

The simple solution is not accurate, but the more elaborate is not either.
Moreover, the simple solution has the benefit of *not* introduting a jump
whenever the pivot changes.

As a result of extended expriments and investigation, my conclusion is
that the only accurate solution would be really to *integrate* all contributions
over all frames. Which we avoid, for sake of responsiveness and interactivity